fix: corregir autenticacion y rediseñar interfaz completa
- Fix auth config para usar modelo Usuario en vez de User
- Crear vista perfil/index.blade.php (faltaba, causaba 500)
- Fix RouteServiceProvider HOME apuntando a /perfil
- Fix withErrors formato en CalcularController
- Consolidar migraciones de resultado en una sola
- Eliminar migracion huérfana de sumas y vistas muertas (home, suma, welcome)

Rediseño profesional dark-theme:
- Layout con Tailwind config, Inter font, paleta brand
- Navbar con dropdown de 4 operaciones y enlace Casino
- Pagina inicio con hero, features y CTA
- Login/Register/Perfil con glass cards y dark inputs
- Calculadora, resultados y casino con diseño consistente
- Footer profesional, CSS limpio sin duplicados
- Añadir Alpine.js collapse plugin para menu movil

+++ database/migrations/2026_07_10_082118_create_operaciones_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
 
             $table->string('tipo');
 
-            $table->integer('resultado');
+            $table->decimal('resultado', 12, 6);
 
             $table->timestamps();
